Move test-utils.js file to the utils folder and rename it to be custom-dimensions.js.
